Works mostly fine but the interface must be expanded
We have to add in the interface the possibility to say what kind of user they are (i.e. student, family, petOwner) and map these values to 1 if true, 0 if false, and -1 if not selected - afterwards we can connect it so to make the update using the hashmap infPrefMap (user type, value). We also have to add some rows in the preference (for like pool and some other stuff that is still missing ) and decide how to present all this stuff in the interface (we can also decide not to show some preference but then they must be initialized to -1 in anycase)
Finally we need to add more houses to the ontology (remember to use this ontology im uploading) and add extra features of reasoning if we want them